move the check cli into its own package and expose it
